+++ locales/english.js
@@ -3177,6 +3177,7 @@ const descriptions = {
     "log finished a quest": "Finished a quest: \"{v1}\"",
     "log finished a quest task": "Finished a task for a quest: \"{v1}\"",
     "log quest progress": "Made some progress in quest: \"{v1}\"",
+    "log ant swarm closes in": "You put your point through the middle of it, and the rest of it arrives anyway - up your sleeves and into every gap you have.",
     "log frog sticky tongue": "The frog's long tongue leaves you covered in sticky saliva!",
     "log frog toxins on hit": "The frog's attack leaves some of its toxins on you!",
     "log frog toxins bare hands": "Touching the frog with your bare hands leaves them covered in toxins!",